git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene3661@1236796 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/java/org/apache/lucene/search/TermStatistics.java
@@ -25,10 +25,10 @@ import org.apache.lucene.util.BytesRef;
  */
 public class TermStatistics {
   private final BytesRef term;
-  private final int docFreq;
+  private final long docFreq;
   private final long totalTermFreq;
   
-  public TermStatistics(BytesRef term, int docFreq, long totalTermFreq) {
+  public TermStatistics(BytesRef term, long docFreq, long totalTermFreq) {
     assert docFreq >= 0;
     assert totalTermFreq == -1 || totalTermFreq >= docFreq; // #positions must be >= #postings
     this.term = term;
@@ -43,7 +43,7 @@ public class TermStatistics {
   
   /** returns the number of documents this term occurs in 
    * @see IndexReader#docFreq(String, BytesRef) */
-  public final int docFreq() {
+  public final long docFreq() {
     return docFreq;
   }
