+++ lucene/src/java/org/apache/lucene/search/similarities/AfterEffectB.java
@@ -27,7 +27,7 @@ public class AfterEffectB extends AfterEffect {
   @Override
   public final float score(BasicStats stats, float tfn) {
     long F = stats.getTotalTermFreq()+1;
-    int n = stats.getDocFreq()+1;
+    long n = stats.getDocFreq()+1;
     return (F + 1) / (n * (tfn + 1));
   }
